+++ src/modules/autoLogin.js
@@ -182,13 +182,6 @@ function initAutoLogin() {
     abortLogin();
   });
 
-  $(".login_button_td").append(
-    '<div id="backToLastPage" style="text-align: center; margin: 23px 0 0 128px"><input type="checkbox" id="backToLastPage_chbx" style="vertical-align: middle"/><label for="backToLastPage_chbx">Az utolsó megtekintett oldalra vissza</label></div>'
-  );
-  $("#backToLastPage_chbx").change(function () {
-    saveLastPageSetting();
-  });
-
   loginTimer = window.setInterval(() => {
     loginCount--;
     submit.attr("value", `${loginButtonText} (${loginCount})`);
@@ -199,20 +192,6 @@ function initAutoLogin() {
       submit.attr("value", `${loginButtonText}...`);
     }
   }, 1000);
-}
-
-//Save the saved state of the last page setting
-function saveLastPageSetting() {
-  storage.set("backToLastPage", utils.getDomain(), document.getElementById("backToLastPage_chbx").checked);
-  storage.set("newLogin", utils.getDomain(), document.getElementById("backToLastPage_chbx").checked);
-}
-
-//Load the saved state of the last page setting
-function loadLastPageSetting() {
-  if (storage.get("backToLastPage", utils.getDomain())) {
-    storage.set("newLogin", utils.getDomain(), true);
-    $("#backToLastPage_chbx").get(0).checked = true;
-  }
 }
 
 // Abort the auto login countdown
@@ -252,6 +231,5 @@ module.exports = {
   initialize: () => {
     initUserSelect();
     initAutoLogin();
-    loadLastPageSetting();
   },
 };
